refactor(desktop): split transcript cards and chat view
